+++ libft/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: ghan <ghan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/03 17:19:50 by ghan              #+#    #+#             */
-/*   Updated: 2021/05/03 17:19:54 by ghan             ###   ########.fr       */
+/*   Updated: 2021/09/22 22:39:05 by ghan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	unsigned int i;
+	size_t	i;
 
 	i = 0;
 	if (!dst && !src)
 		return (dst);
 	while (i < n)
 	{
-		*(unsigned char*)(dst + i) = *(unsigned char*)(src + i);
+		*(unsigned char *)(dst + i) = *(unsigned char *)(src + i);
 		i++;
 	}
 	return (dst);
