+++ libft/ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: ghan <ghan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/03 17:36:44 by ghan              #+#    #+#             */
-/*   Updated: 2021/05/03 17:36:46 by ghan             ###   ########.fr       */
+/*   Updated: 2021/09/22 22:41:32 by ghan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	unsigned int i;
+	size_t	i;
 
 	if (len == 0)
 		return (dst);
@@ -23,7 +23,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 		i = 0;
 		while (i < len)
 		{
-			*(unsigned char*)(dst + i) = *(unsigned char*)(src + i);
+			*(unsigned char *)(dst + i) = *(unsigned char *)(src + i);
 			i++;
 		}
 	}
@@ -32,10 +32,10 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 		len--;
 		while (len > 0)
 		{
-			*(unsigned char*)(dst + len) = *(unsigned char*)(src + len);
+			*(unsigned char *)(dst + len) = *(unsigned char *)(src + len);
 			len--;
 		}
-		*(unsigned char*)dst = *(unsigned char*)src;
+		*(unsigned char *)dst = *(unsigned char *)src;
 	}
 	return (dst);
 }
