+++ libft/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: ghan <ghan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/03 20:26:29 by ghan              #+#    #+#             */
-/*   Updated: 2021/05/03 20:26:31 by ghan             ###   ########.fr       */
+/*   Updated: 2021/09/22 22:31:18 by ghan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	src_l;
+	size_t	i;
+	size_t	src_l;
 
 	src_l = 0;
 	while (src[src_l])
